+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Version changes are pinned to SDK releases.
 
 ## [Unreleased]
 
-- client: Insurance fund functionality, whitelist checks, deposit & withdrawal. ([#9](https://github.com/zetamarkets/sdk/pull/9))
+- error: Move error related functionality to error.ts. ([#6](https://github.com/zetamarkets/sdk/pull/6))
+- insurance-client: Insurance fund functionality, whitelist checks, deposit & withdrawal. ([#9](https://github.com/zetamarkets/sdk/pull/9))
 - exchange: Insurance functionality, whitelist a user, rebalance vaults. ([#9](https://github.com/zetamarkets/sdk/pull/9))
 
 ## [0.7.3] 2021-11-29